server/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
MarkupSafe==3.0.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyotp==2.9.0
qrcode==8.0
setuptools==75.6.0
sqlparse==0.5.3
typing_extensions==4.12.2
Werkzeug==3.1.3
